CIV: Move reboot and boot tests to a separate test suite file
https://issues.redhat.com/browse/CLOUDX-241
